Move more legacy stuff under ./bin/legacy
