+++ bin/legacy/commands/ci.js
@@ -1,4 +1,4 @@
-const { builder, handler } = require('../cml/repository/configure');
+const { builder, handler } = require('../../cml/repository/configure');
 
 exports.command = 'ci';
 exports.description = false;
