+++ bin/legacy/commands/send-comment.js
@@ -1,4 +1,4 @@
-const { builder, handler } = require('../cml/report/create');
+const { builder, handler } = require('../../cml/report/create');
 
 exports.command = 'send-comment <markdown file>';
 exports.description = false;
