+++ bin/legacy/commands/send-github-check.js
@@ -1,4 +1,4 @@
-const { builder, handler } = require('../cml/check/create');
+const { builder, handler } = require('../../cml/check/create');
 
 exports.command = 'send-github-check <markdown file>';
 exports.description = false;
