+++ bin/legacy/link.js
